[precompile] Support external data for serialization.
Under some corner cases, users don't directly pass us a free function
or module.forward. Instead, there're multiple levels of wrappers and
captured closures applied to module.forward, so we end up serialzing quite
a few captured data from the compiled function's closure. This is not great
when things like torch.nn.Module end up being captured in function's closure,
and we shouldn't serialize the whole module in this case.

To solve this issue, we specially handle the common cases when the following
objects are being saved:
1. torch.nn.Module
2. Nested functions.

When these data are being serialized, we will mark these as "external data"
since they may contain unserialzable and untrackable states which is
usually easier to be maintained from user side. So the a call to save_compiled_function()
looks like the following:
```
result: AOTCompileSaveResult = compiled_fn.save_compiled_function("/path")

```

On the loading side, user are supposed to provide the external data dict. If any key
is missing, load_compiled_function will throw an error.
```
torch.compiler.load_compiled_function(f, external_data={...})
```

Note that it's not ideal for user to compile a typical function like this, so this
is only meant for power users for whom rewriting model is harder than maintaining
this data map.

ghstack-source-id: 39110c3bd9a6a7b5a0333abd53d54ccb7ec6e7a1
Pull Request resolved: https://github.com/pytorch/pytorch/pull/170846