[precompile] Support serializing nested function in the context of guard preservation.
Nested functions shouldn't be serialized on the surface, but actually are quite popular
objects to be serialiezd when they capture some tensor objects which produces
tensor guards. Since dynamo constructs guards based on the object tree, we are
required to preserve the root object (nested function) before we serialize
tensor metadata in guard serialization.

This also allows us to remove a previous hack `source_get_cache` which
skips things like nested function in hierachy but ultimately not so great to support
in the long run. After some discussion with Animesh we decided to remove it but
that is blocked by failed test like `test_closure_var_missing`. Now by serializing nested
function properly we can safely remove `source_get_cache` from the codebase.

Test Plan:

Note that we don't need to add a new test because the following should
cover this, and this test previously works due to a hack we're also
removing in the current PR:

pytest test/dynamo/test_guard_serialization.py -k closure

ghstack-source-id: 1aacceacdb8f36b0b7341d3c9c2ef465ac0c17ac
Pull Request resolved: https://github.com/pytorch/pytorch/pull/171156